pipeline/requirements.txt
# Snowflake
snowflake-connector-python[pandas]>=3.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0

# Utilities
python-dotenv>=1.0.0
